+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-02-15
+## [Unreleased] - 2022-02-17
 
 ### Added
 
@@ -19,6 +19,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### CI/CD
 
 ### Other
+
+
+## [3.1.1] - 2022-02-17
+
+### Fixed
+* Fixed inefficient memory usage in KendallTauDistance (issue #183)
+* Fixed inefficient memory usage in KendallTauSequenceDistance (issue #184)
+* Fixed inefficient memory usage in WeightedKendallTauDistance (issue #185)
+* Fixed inefficiency in WeightedKendallTauDistance related to redundant computation (issue #182)
 
 
 ## [3.1.0] - 2022-02-15